\name{ssgraph-package}
\alias{ssgraph-package}
\docType{package}

\title{ Bayesian Graphical Estimation using Spike-and-Slab Priors }

\description{
The \code{R} package \pkg{ssgraph} provides statistical tools for Bayesian estimation in graphical models.       
}

\author{Reza Mohammadi <a.mohammadi@uva.nl>}

\keyword{ package }
